+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+## Changed
+
+- Depends on `lumberjack` 2.0 or greater
+- You can now use `level` and `severity` interchangeably when checking for log entry attributes. "Severity" is more correct and is the preferred term.
+
 ## 1.2.2
 
 ### Changed